refactor(client): improve vars naming
