+++ client/src/components/input/index.ts
@@ -1,3 +1,3 @@
-import { Input } from 'components/input/Input';
+import { Input } from 'components/Input/Input';
 
 export default Input;
