+++ client/src/graphql/sign-up/signup.mutation.ts
@@ -1,4 +1,4 @@
-import { gql } from "@apollo/client";
+import { gql } from '@apollo/client';
 
 export const SIGN_UP_MUTATION = gql`
 	mutation SignUp($username: String!, $email: String!, $password: String!) {
