+++ client/src/helpers/getFormErrors.ts
@@ -13,8 +13,8 @@ interface InputsValueArg {
 	confirmedPassword: string;
 }
 
-export const getInputErrors = (inputsValue: InputsValueArg) => {
-	const { username, email, password, confirmedPassword } = inputsValue;
+export const getFormErrors = (fieldsValue: InputsValueArg) => {
+	const { username, email, password, confirmedPassword } = fieldsValue;
 	const errorMessages: { [key: string]: string } = {};
 
 	if (!username.match(USERNAME_REGEX)) errorMessages.username = USERNAME_MUST_BE_LONG_ERROR_MESSAGE;
